Clarify usage comment for set() functions (#67)
It was not immediately clear that supplying a value of false for the "exists" argument has a special meaning. It causes the "NX" argument to be sent to Redis, as opposed to no argument when supplying a value of nil. At least that was my experience when using this function for the first time - the implications of true/false/nil were not explained until I read the source code for this function.